pyproject.toml
[project]
name = "autolessonregistration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.60.0",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.9",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.9",
    "requests>=2.32.5",
    "rq>=2.6.1",
    "selenium>=4.40.0",
    "webdriver-manager>=4.0.2",
]
